avoid colourisation in dumb terminals
Dumb terminals, identified via TERM=dumb, can only interpret a
limited number of escape sequences.  In particular, they do not
support colourisation or moving the cursor.

This patch enables use of phlay in dumb terminals such as rc(1)
from Plan 9 and Emacs' built-in terminal.

The patch can be tested this way:

	TERM=dumb ./phlay

The output should not contain any colours when --color is set to auto
(default).