Fix version number comparison
`"0.1.10" > "0.1.9"` is `False`, which results in phlay being eternally
stuck at an older version. Fix this by using a version comparator
instead of a string comparison.

Note that the StrictVersion class only support version numbers of at
most three parts. "1", "1.2", "1.2.3" are valid, but "1.2.3.4" is not.

To make sure that users don't get stuck on 0.1.9 or earlier, the
package version should be bumped to 0.2.0.