Set parent revision if the parent commit has a revision and is for the same bug
A typical case is to have multiple commits for a single bug, and when
some of them need updates after review, one may want to update only
those revisions rather than the whole stack.

When doing so, parent information is then lost. With this change, the
parent information is preserved, if the commit that is parent of the
start of the processed range is for the same bug and has a revision
in its commit message.