Support setting diff description via --comment
The diff description is the "Description" column of the diff history
view in Phabricator. Before this commit, that field was always empty.
In contrast, `arc` (Arcanist, the official tool from Phabricator) did
allow the field to be set, defaulting to "Revision updated.".

This commit introduces support for setting the "Description" field of
the history view, via the `--comment` flag.

When `phlay --comment='rebased on XXX'` is used, the effect is the same
as before, with the following differences:

- The "Description" column at the History view shows "rebased on XXX"
- The feed at the revision shows "rebased on XXX" under the heading
  `[author] updated this revision to Diff [diffid].`
- If there are more changes (e.g. title update), then it is followed by
  `[author] updated this revision to Diff [diffid]` with those changes.