Add a way to specify non-blocking reviewers.
Sometime I tag a reviewer group for review, and one other person which isn't in
that reviewer group but that could review that patch for example, or multiple
people that could review the patch, but for which I don't need review from each
of them.

This allows to add non-blocking reviewers for that kind of stuff.